% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{combine_fore_err_corr}
\alias{combine_fore_err_corr}
\title{Combine COVID19 and Influenza Forecasts}
\usage{
combine_fore_err_corr(
  prof_data = NULL,
  dates_frcst_list = NULL,
  simdat_list = NULL,
  nfrcst = NULL,
  err_corr = 0,
  method_name = "semi_sorted_randA"
)
}
\arguments{
\item{prof_data}{- the complete data structure}

\item{dates_frcst_list}{- a list of length two with dates (fit and forecast) for each
pathogen}

\item{simdat_list}{- a list of length two with an array of trajectories for each pathogen}

\item{nfrcst}{integer - the number of days/weeks that were simulated beyond
the fitting time period.}

\item{method_name}{character - 'semi_sorted_randA' or 'lin_scale'. 'semi_sorted_randA'
is a more proper combination of profiles.  'lin_scale' uses the uncertainty
range of 'semi_sorted_randA', but ensures that the combined forecast has a
median equal to the sum of the aggregate forecast medians.}

\item{cor_val}{numeric [-1, 1] designating the expected error/uncertainty
correlation between the aggregate forecasts.}
}
\value{
a list with two items
simdat_both - a list of length two with the two estimates for the combined burden
(random and ordered)
inc_both - the combined reported hospitalization
dates_both - dates array for combined forecast
}
\description{
Forecasts can be combined using two procedures using the error-correlation:
semi-random trajectory matching 'semi_sorted_randA' or a median-preserving
method that linearly scales uncertainty across known anchor points 'lin_scale'.
}
