% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download_format.R
\name{csv_to_prof}
\alias{csv_to_prof}
\title{TODO: UPDATE DOCUMENTATION}
\usage{
csv_to_prof(
  filepath,
  population,
  location = "NA",
  fit_start = NULL,
  fit_end = NULL
)
}
\description{
TODO: UPDATE DOCUMENTATION
}
