% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dates_to_int}
\alias{dates_to_int}
\title{Date to Integer conversion}
\usage{
dates_to_int(dates)
}
\arguments{
\item{dates}{array of dates for incidence}
}
\value{
times - an array of day numbers
}
\description{
Create a days number array from a dates array
}
\keyword{internal}
