% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{dmeas}
\alias{dmeas}
\title{Log-Likelihood calculation}
\format{
An object of class \code{Csnippet} of length 1.
}
\usage{
dmeas
}
\value{
LLK
}
\description{
Log-Likelihood calculation
}
\keyword{datasets}
