% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fore_agrgt_funs.R
\name{eval_comb}
\alias{eval_comb}
\title{Evaluate the combined uncertainty of two forecasts aggregated by addition.}
\usage{
eval_comb(method_name, A, B, cor_val, sample_var = NULL)
}
\arguments{
\item{method_name}{character - 'semi_sorted_randA' or 'lin_scale'.}

\item{A}{numeric vector of sorted forecast profiles for pathogen A.}

\item{B}{numeric vector of sorted forecast profiles for pathogen B.}

\item{cor_val}{numeric [-1, 1] designating the expected error/uncertainty correlation between the forecasts for A and B.}

\item{sample_var}{numeric non-negative error correlation variance. Used only for 'semi_sorted_randA', this controls the variance in the profile mixing matrix.}
}
\value{
numeric vector of the combined A and B profiles/samples.
}
\description{
This function is a wrapper for multiple methods ('semi_sorted_randA', 'lin_scale'). 'semi_sorted_randA' is a more proper combination of profiles.  'lin_scale' uses the uncertainty range of 'semi_sorted_randA', but ensures that the combined forecast has a median equal to the sum of the aggregate forecast medians.  The methods assume that vectors A and B are sorted samples of a single forecast target for pathogen A and pathogen B.
}
