% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download_format.R
\name{fetch_hhs_last_modified}
\alias{fetch_hhs_last_modified}
\title{Fetch the date (in GMT) that the `g62h-syeh` dataset was last modified, as
reported by healthdata.gov}
\usage{
fetch_hhs_last_modified()
}
\value{
POSIXct Date. Date (where tz=GMT) that the `g62h-syeh` dataset was
        last modified. If the GET response code is not 200, NULL is returned.
}
\description{
Fetch the date (in GMT) that the `g62h-syeh` dataset was last modified, as
reported by healthdata.gov
}
\examples{
filepath <- "HHS_daily-hosp_state__210226120000.csv"
posix_timestamp <- filename_timestamp_to_posix(filepath)
last_modified <- fetch_hhs_last_modified()
posix_timestamp < last_modified

}
