% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download_format.R
\name{fetch_season_tstart_tend}
\alias{fetch_season_tstart_tend}
\title{Fetch the start and end dates for a given `season`.}
\usage{
fetch_season_tstart_tend(season)
}
\arguments{
\item{season}{integer. The requested season.}
}
\value{
list of Dates. Contains elements `t0` and `t1` corresponding to
        the season's start date and end date, respectively.
}
\description{
Fetch the start and end dates for a given `season`.
}
\examples{
date_range <- fetch_season_tstart_tend(2023)

}
