% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download_format.R
\name{filename_timestamp_to_posix}
\alias{filename_timestamp_to_posix}
\title{Convert POSIX timestamp from filename to POSIXct object.}
\usage{
filename_timestamp_to_posix(filepath)
}
\arguments{
\item{filepath}{Character string. The filepath containing the filename with the POSIX timestamp.}
}
\value{
A POSIXct object representing the timestamp extracted from the filename.
}
\description{
This function extracts a POSIX timestamp from a filename following a specific pattern,
then converts it to a POSIXct object.
}
\examples{
filepath <- "HHS_daily-hosp_state__210226120000.csv"
posix_timestamp <- filename_timestamp_to_posix(filepath)
posix_timestamp

}
