% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download_format.R
\name{format_hhs_state}
\alias{format_hhs_state}
\title{Format HHS hospitalization data for fitting.}
\usage{
format_hhs_state(state_data, fit_col, inc_type, loc_name, pop, disease)
}
\arguments{
\item{state_data}{Dataframe. The output from get_HHS_state().}

\item{fit_col}{character. The name of the incidence columnn in 'state_data'
that should be used for model fitting.}

\item{pop}{integer.  The location's population.}

\item{disease}{character.  The name of the disease.}

\item{state}{character.  The name of the location.}
}
\value{
A list of lists.
    list(influenza=list(disease="influenza",
                        population=12345,
                        loc_name="profville",
                        inc_type="hosp_admits",
                        data=data.frame(date=dates,
                                      inc=incidence)))
}
\description{
After subsetting the admissions data to a single location, now format for
fitting.
}
\examples{
format_hhs_state(state_data, fit_col, inc_type,loc_name, pop, disease)

}
