% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{fromEst}
\alias{fromEst}
\title{Transform variables back}
\format{
An object of class \code{Csnippet} of length 1.
}
\usage{
fromEst
}
\value{
variables transformed back
}
\description{
Transform variables back
}
\keyword{datasets}
