% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download_format.R
\name{get_HHS_state}
\alias{get_HHS_state}
\title{Subset HHS PROTECT hospitalizations by state, metric, and dates.}
\usage{
get_HHS_state(
  hosp_data = NULL,
  get_cols = c("previous_day_admission_adult_covid_confirmed",
    "previous_day_admission_adult_covid_confirmed_coverage",
    "previous_day_admission_pediatric_covid_confirmed",
    "previous_day_admission_pediatric_covid_confirmed_coverage", "deaths_covid",
    "deaths_covid_coverage"),
  state = "CA",
  start_date = NULL,
  end_date = NULL,
  shift_dates = F
)
}
\arguments{
\item{hosp_data}{dataframe. Full HHS PROTECT dataset.}

\item{get_cols}{vector of strings. The names of the HHS PROTECT columns to
be extracted.}

\item{state}{character vector (length=2). Abbreviation of the state to be
subset.}

\item{start_date}{Date. Start date (inclusive) for data subsetting.}

\item{end_date}{Date. End date (inclusive) for data subsetting.}

\item{shift_dates}{logical (FALSE). Previous versions of the HHS data listed
admissions by day reported, which required a one-day shift to be
'day of admission'.}
}
\value{
Dataframe. This has columns 'state', 'data', and everything specified
in get_cols. Rows are sorted by date. The dataframe contains only data
for the specified state.
}
\description{
Subset HHS PROTECT hospitalizations by state, metric, and dates.
}
