% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download_format.R
\name{get_loc_pop}
\alias{get_loc_pop}
\title{Retrieve population of a U.S. state or territory.}
\usage{
get_loc_pop(location = "US")
}
\arguments{
\item{location}{character. Intended to be a two-letter abbreviation, but will
also attempt to match to location full names.}
}
\value{
integer. Population of the specified location.
}
\description{
Retrieve population of a U.S. state or territory.
}
\examples{
get_loc_pop("CA")

}
