% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download_format.R
\name{hhs_2_PROF}
\alias{hhs_2_PROF}
\title{Convert HHS hospitalization data to PROF format.}
\usage{
hhs_2_PROF(hhs_path = NULL, season = NULL, state = NULL)
}
\arguments{
\item{hhs_path}{character. The local path to HHS data file.}

\item{season}{integer. The start year (YYYY) of season to model.}

\item{state}{character. 2-letter state abbreviation.}
}
\value{
PROF data structure (list of pathogen-data)
}
\description{
This function converts HHS hospitalization data to the structure that
works with with PROF fitting routines.  This function retrieves data 
for an entire season and expects the HHS data is already downloaded. 
Before sending to fitting procedures, the output of this function
needs fitting date ranges (see PROF::hhs_set_fitdates()).
}
