% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download_format.R
\name{hhs_hosp_state_down}
\alias{hhs_hosp_state_down}
\title{Download daily state-level HHS PROTECT hospitalization admission data
to a CSV.}
\usage{
hhs_hosp_state_down(down_dir = "~/", down_filename = NULL)
}
\arguments{
\item{down_dir}{character string. The directory path to download to.}

\item{down_filename}{character string. The filename to download to.}
}
\value{
an integer. 0 for success and non-zero for failure.
}
\description{
Download daily state-level HHS PROTECT hospitalization admission data
to a CSV.
}
