% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download_format.R
\name{hhs_set_fitdates}
\alias{hhs_set_fitdates}
\title{Add fit-data structures to PROF_data}
\usage{
hhs_set_fitdates(prof_data = NULL, fit_start = NULL, fit_end = NULL)
}
\arguments{
\item{prof_data}{list. Generally the output from PROF::hhs_2_PROF().}

\item{fit_start}{Date. The lower limit (inclusive) of the data range to
be fit.}

\item{fit_end}{Date. The upper limit (inclusive) of the data range to
be fit.}
}
\value{
PROF_data structure that now includes a $data_fit entry for 
each pathogen.
}
\description{
Fit-data structures explicitly state the data to be passed to the fitting
routine.  This allows the user to exclude data from the begining or end 
of the season in the model fit.
}
