% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download_format.R
\name{hhs_set_fitdates_stat}
\alias{hhs_set_fitdates_stat}
\title{Add fit-data-stat structures to PROF_data}
\usage{
hhs_set_fitdates_stat(prof_data = NULL, fit_start = NULL, fit_end = NULL)
}
\arguments{
\item{prof_data}{list. Generally the output from PROF::hhs_2_PROF().}

\item{fit_start}{list. Each element is a Date or NULL with the lower limit
(inclusive) of the data range to be fit.}

\item{fit_end}{Date. The upper limit (inclusive) of the data range to
be fit.}
}
\value{
PROF_data structure that now includes a $data_fit_stat entry for
each pathogen.
}
\description{
Fit-data-stat structures explicitly state the data to be passed to the statistical
fitting routine.  This allows the user to exclude data from the beginning or end
of the season in the model fit.
}
