% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{init}
\alias{init}
\title{Initial Conditions SIRH Model}
\format{
An object of class \code{Csnippet} of length 1.
}
\usage{
init
}
\value{
Initial values of all compartments
}
\description{
Initial Conditions SIRH Model
}
\keyword{datasets}
