% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{init.seirh}
\alias{init.seirh}
\title{Initial Conditions SEIRH Model}
\format{
An object of class \code{Csnippet} of length 1.
}
\usage{
init.seirh
}
\value{
Initial values of all compartments
}
\description{
Initial Conditions SEIRH Model
}
\keyword{datasets}
