% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_struct.R
\name{init_model_seirh}
\alias{init_model_seirh}
\title{Produce an empty parameter list for the SEIRH model.}
\usage{
init_model_seirh()
}
\value{
A list of lists for modeling and fitting parameters
}
\description{
Each pathogen to be fit/forecast requires a set of model and fit parameters.
This function produces an empty list of lists with the structure and entries
required by the fitting function(s) for an SEIRH model.
}
