% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_struct.R
\name{init_par_list}
\alias{init_par_list}
\title{Full parameter list structure for multiple diseases.}
\usage{
init_par_list(
  diseases = c("covid19", "influenza"),
  models = c("seirh", "sirh")
)
}
\arguments{
\item{diseases}{Character vector listing all diseases to be fit.}

\item{models}{Character vector stating the model to be used for each
disease.}
}
\value{
A list of disease parameter lists.
}
\description{
Each disease to be fit must have an entry in the parameter list. This
function simply loops through pathogens and populates its list entry
with the appropriate model structure.
}
