% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{init_param}
\alias{init_param}
\title{Set disease specific initial guess for some model parameters}
\usage{
init_param(mymodel, inc)
}
\arguments{
\item{mymodel}{model name sirh or seirh}

\item{inc}{time-series of hospitalization}
}
\value{
param a model specific list with initial guess for parameters:

gamma - generation time

mu_H1H2 - holding time between sub-compartments 1 and 2 of H

Beta - force of infection

pH - hospitalization probability

rho - reporting rate (0.95)

mu_EI - latency period (seirh only)
}
\description{
Based on the disease set the initial guess for model parameters
}
\keyword{internal}
