% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download_format.R
\name{load_HHS_csv}
\alias{load_HHS_csv}
\title{}
\usage{
load_HHS_csv(hhs_file = NULL)
}
\arguments{
\item{hhs_file}{Character string. Path to CSV file.}
}
\value{
dataframe of all HHS PROTECT columns
}
\description{
api_url = "https://healthdata.gov/api/views/g62h-syeh/rows.csv?accessType=DOWNLOAD"
}
\details{

}
\examples{
load_HHS_csv(hhs_file = NULL)

}
