% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROF-package.R
\docType{data}
\name{loc_pops}
\alias{loc_pops}
\title{U.S. location fips and populations info.}
\format{
A data frame with 54 rows and 4 variables
}
\usage{
data(loc_pops)
}
\description{
A dataset containing the FIPS codes and populations of 53 states and
 territories as well as the national population. The variables are as follows:
}
\details{
\itemize{
  \item abbreviation. Two-letter location abbreviation
  \item location. Two-digit location FIPS code
  \item location_name. Full name of location
  \item population. Integer estimate of location population
}
}
\keyword{datasets}
