% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mixed_forecast.R
\name{plot_mixed_forecast}
\alias{plot_mixed_forecast}
\title{Create a Mixed Forecast}
\usage{
plot_mixed_forecast(prof_data, forecast_list, forecast_stat_list)
}
\arguments{
\item{prof_data}{data structure for diseases}

\item{forecast_list}{the list output of plot_forecast function}

\item{forecast_stat_list}{the list output of plot_stat_forecast function}
}
\value{
arrange_plot - a list of plot(s)
total - data frame(s) used for the plots
forecast_mixed - a list with the mixed forecast for each pathogen
}
\description{
Use the mechanistic and statistical forecasts to create a mixed forecast
Forecast are combined for each pathogen separately
}
