% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plot_prof_data}
\alias{plot_prof_data}
\title{Basic Plot Incidence}
\usage{
plot_prof_data(prof_data, filename = NULL)
}
\arguments{
\item{prof_data}{- the complete data structure}

\item{filename}{- if NULL print to screen else save to PDF}
}
\value{
plot object to screen or to file
}
\description{
Basic plotting function
}
