% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stat_fit.R
\name{plot_stat_fit}
\alias{plot_stat_fit}
\title{Plots the Results of a Baseline Statistical Model to the dataset(s)}
\usage{
plot_stat_fit(prof_data, ntraj = NULL, filename = NULL)
}
\arguments{
\item{prof_data}{data structure for diseases
for each disease it includes:
disease - covid19 or influenza
population - population size
loc_name - location name
inc_type - incidence type, e.g. hosp_admits
data - all available data as a 2D structure of dates and incidence
data_fit - subset of data for fitting (can be eqaul to data)
 @param ntraj - integer number of stochastic trajectories, default 1000
 @filename - if NULL print plot to screen, if not save also to filename}
}
\value{
plots to screen and file and a list of two items:
stat_fit_traj - a list with a list for each disease: statistical trajectories,
dates, and reported incidence.

stat_pl - a list of statistical fit plots
}
\description{
Plots the Results of a Baseline Statistical Model to the dataset(s)
}
