% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stat_forecast.R
\name{plot_stat_forecast}
\alias{plot_stat_forecast}
\title{Baseline statistical forecast}
\usage{
plot_stat_forecast(prof_data, ntraj = NULL, nfrcst = NULL, filename = NULL)
}
\arguments{
\item{prof_data}{data structure for diseases
for each disease it includes:
disease - covid19 or influenza
population - population size
loc_name - location name
inc_type - incidence type, e.g. hosp_admits
data - all available data as a 2D structure of dates and incidence
data_fit - subset of data for fitting (can be eqaul to data)
 @param ntraj - integer number of stochastic trajectories, default 1000
 @param nfrcst - number of time units to produce a forecast for (assume to be the same
 cadence as the data), default is 35 days
 @filename - if NULL print plot to screen, if not save also to filename}
}
\value{
plots to screen and  file
forecast_traj - a list with a list for each disease and for the combined burden
(random and ordered): trajectories,
dates, and reported incidence.
total_list - a list with the data frames used for each of the plots, these include the median and quantiles
wis_df - a list of data frame with the WIS score of the forecast if available.  If not it contains a NULL
}
\description{
Calculates and plots a baseline statistical forecast for to the dataset(s)
}
