% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wis.R
\name{plot_wis}
\alias{plot_wis}
\title{plot_wis}
\usage{
plot_wis(wis_data, loc = NA, filename = NULL)
}
\arguments{
\item{wis_data}{- a list with the WIS score (or NULL) for each forecasted pathogen}

\item{loc}{string location name}

\item{filename}{optional for saving plot(s) to a file}
}
\value{
A data frame with all WIS information
}
\description{
Plot the WIS score of Forecasts
}
