% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROF-package.R
\docType{package}
\name{PROF-package}
\alias{PROF}
\alias{PROF-package}
\title{PROF: Package for Respiratory disease Open-source Forecasting}
\description{
A Package for fitting publicaly available (or user provided) confirmed hospital admission data and providing 1-4 weeks ahead probabilistic forecasts. PROF supports two respiratory viruses: influenza and COVID-19 and when requested will provide the combined burden of both.
}
\details{
Additional details...
}
\keyword{internal}
