% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_struct.R
\name{read_par_list_yaml}
\alias{read_par_list_yaml}
\title{Read parameter list from YAML file.}
\usage{
read_par_list_yaml(file_path = NULL)
}
\arguments{
\item{file_path}{Full path for file to be read.}
}
\value{
List of parameter lists
}
\description{
The fitting function requires a specific list structure for the parameters.
See PROF/parameters/param_exmpl.yml for an example YAML file with 
documentation. Or use PROF::init_pars_list() to generate an appropriate 
structure in R.
}
