% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{rmeas}
\alias{rmeas}
\title{Given the number in the H compartment the reporting
probability and and baseline value calculate the
number of cases using Poisson sampling}
\format{
An object of class \code{Csnippet} of length 1.
}
\usage{
rmeas
}
\value{
cases, integer
}
\description{
Given the number in the H compartment the reporting
probability and and baseline value calculate the
number of cases using Poisson sampling
}
\keyword{datasets}
