% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{score_forecast}
\alias{score_forecast}
\title{Weight-Interval-Score for forecast}
\usage{
score_forecast(obs = obs, simdat = NULL)
}
\arguments{
\item{:}{simdat - realization matrix - used to calculate quantiles of forecast}
}
\value{
wis_arr - An array of WIS scores with the same length as observations
}
\description{
A wrapper for calls to scoringutils package giving WIS score to forecasts
}
\examples{
score_forecast(obs= obs, simdat = simdat)
}
