% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{seirh.ode}
\alias{seirh.ode}
\title{Deterministic SEIRH Euler Time Step}
\format{
An object of class \code{Csnippet} of length 1.
}
\usage{
seirh.ode
}
\value{
change in each compartment, real
}
\description{
Advance SEIRH model by one deterministic time step
}
\keyword{datasets}
