% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{seirh.step}
\alias{seirh.step}
\title{Stochastic SEIRH Euler Time Step}
\format{
An object of class \code{character} of length 1.
}
\usage{
seirh.step
}
\value{
change in each compartment, integer
}
\description{
Advance SEIRH model by one stochastic time step
}
\keyword{datasets}
