% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fore_agrgt_funs.R
\name{semi_sorted_randA}
\alias{semi_sorted_randA}
\title{Generate a trajectory-matching matrix for the 'semi_sorted_randA' forecast 
aggregation method.}
\usage{
semi_sorted_randA(
  n,
  correlation_val = 1,
  sample_var = n - n * abs(correlation_val)
)
}
\arguments{
\item{n}{integer - the number of forecast profiles}

\item{correlation_val}{numeric [-1, 1] - a continuous parameter that defines
a correlation between the aggregate uncertainties. NOTE: only the sign of 
correlation_val is used here.  correlation_val and sample_var are not
independent (see sample_var below).}

\item{sample_var}{numeric [0, n] - Sample var controls the strength of 
correlation in the mixing matrix.  We usually define it as 
(n - n*abs(correlation_val)),
but it is an independent input in this function so the user may define their 
own relationship.}
}
\description{
Generate a trajectory-matching matrix for the 'semi_sorted_randA' forecast 
aggregation method.
}
