% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_init_states}
\alias{set_init_states}
\title{Set initial values for model states - SIRH or SEIRH}
\usage{
set_init_states(pop, I0, mymodel)
}
\arguments{
\item{pop}{population size, integer}

\item{I0}{estimate for initial number of cases}

\item{mymodel}{model name sirh or seirh}
}
\value{
init_state named list with initial number of individuals in each compartment:

S0 - susceptible

E0 - exposed (seirh only)

I0 - infectious

R0 - recovered

H0 - hospitalized
}
\description{
Based on the population size and the initial number of cases initialize the
number in each compartment
}
\keyword{internal}
