% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_td_beta}
\alias{set_td_beta}
\title{Set initial values for time-dependent Force of Infection}
\usage{
set_td_beta(nb, ntimes, Beta)
}
\arguments{
\item{nb}{integer number of unique values FOI can have}

\item{ntimes}{integer length of time series}

\item{Beta}{real single initial guess value for the FOI}
}
\value{
td_beta a list with two items:
betac - array with initial guesses for beta
tcng. - array with initial guesses for time of change in beta
}
\description{
Set the values for the force of infection and the time(s) change takes place
}
\keyword{internal}
