% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{sir.ode}
\alias{sir.ode}
\title{Deterministic SIR Euler Time Step}
\format{
An object of class \code{Csnippet} of length 1.
}
\usage{
sir.ode
}
\value{
change in each compartment, real
}
\description{
Advance SIR model by one deterministic time step
}
\keyword{datasets}
