% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{sir.step}
\alias{sir.step}
\title{Stochastic SIR Euler Time Step}
\format{
An object of class \code{character} of length 1.
}
\usage{
sir.step
}
\value{
change in each compartment, integer
}
\description{
Advance SIR model by one stochastic time step
}
\keyword{datasets}
