% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{sirh.ode}
\alias{sirh.ode}
\title{Deterministic SIRH Euler Time Step}
\format{
An object of class \code{Csnippet} of length 1.
}
\usage{
sirh.ode
}
\value{
change in each compartment, real
}
\description{
Advance SIRH model by one deterministic time step
}
\keyword{datasets}
