% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{sirh.step}
\alias{sirh.step}
\title{Stochastic SIRH Euler Time Step}
\format{
An object of class \code{character} of length 1.
}
\usage{
sirh.step
}
\value{
change in each compartment, real
}
\description{
Advance SIRH model by one stochastic time step
}
\keyword{datasets}
