% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{stat_forecast}
\alias{stat_forecast}
\title{Forecast using a baseline statistical model fitted to the data}
\usage{
stat_forecast(data, ntraj, nfrcst)
}
\arguments{
\item{data}{entire data structure of the pathogen}

\item{ntraj}{integer. Number of requested trajectories}

\item{nfrcst}{integer. Number of requested forecast horizons}
}
\value{
simdat a 2D array of dimension (ntraj, (ntimes+nfrcst))
}
\description{
See: file:///Users/michal/Downloads/pnas.2113561119.sapp.pdf for more details (start at bottom of page 3)
This baseline model forecasts a predictive median incidence equal to the incidence in the
most recent data point(s), with uncertainty around the median based on changes in incidence
that were observed in the past of the time series.
Adjustments are made for daily data (e.g., we use the last seven data points for the forecast)
Note that we do not force the median of the forecast forecast to be equal to
the last observed value since the data is likely to be daily.  But we do truncate the distribution
so that it has no negative values.
}
