% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{td.init}
\alias{td.init}
\title{Initial Conditions TD-SIRH Model}
\format{
An object of class \code{Csnippet} of length 1.
}
\usage{
td.init
}
\value{
Initial values of all compartments
}
\description{
Initial Conditions TD-SIRH Model
}
\keyword{datasets}
