% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{td.sirh.ode}
\alias{td.sirh.ode}
\title{Deterministic TD-SIRH Euler Time Step}
\format{
An object of class \code{Csnippet} of length 1.
}
\usage{
td.sirh.ode
}
\value{
change in each compartment, real
}
\description{
Advance TD-SIRH model by one deterministic time step
}
\keyword{datasets}
