% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{td.sirh.step}
\alias{td.sirh.step}
\title{Stochastic TD-SIRH Euler Time Step}
\format{
An object of class \code{character} of length 1.
}
\usage{
td.sirh.step
}
\value{
change in each compartment, real
}
\description{
Advance TD-SIRH model by one stochastic time step
}
\keyword{datasets}
