% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{td3.seirh.ode}
\alias{td3.seirh.ode}
\title{Deterministic TD-SEIRH Euler Time Step}
\format{
An object of class \code{Csnippet} of length 1.
}
\usage{
td3.seirh.ode
}
\value{
change in each compartment, real
}
\description{
Advance TD-SEIRH model by one deterministic time step
This version is for a three value FOI
}
\keyword{datasets}
