% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{td3.seirh.step}
\alias{td3.seirh.step}
\title{Stochastic TD-SEIRH Euler Time Step}
\format{
An object of class \code{character} of length 1.
}
\usage{
td3.seirh.step
}
\value{
change in each compartment, integer
}
\description{
Advance TD-SEIRH model by one stochastic time step
This code is for a 3 value FOI
}
\keyword{datasets}
