% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{td3_seirh_dynamics}
\alias{td3_seirh_dynamics}
\title{Deterministic ODEs for SEIRH model with a
3-value FOI}
\usage{
td3_seirh_dynamics(t, y, parms)
}
\arguments{
\item{t}{- array of times}

\item{y}{- array of states}

\item{parms}{- array of paramters inclucing 'wl'}
}
\value{
res - an array with the derivatives of y
}
\description{
Deterministic ODEs for SEIRH model with a
3-value FOI
}
