% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{toEst}
\alias{toEst}
\title{Transform variables before optimization}
\format{
An object of class \code{Csnippet} of length 1.
}
\usage{
toEst
}
\value{
transformed variables
}
\description{
Transform variables before optimization
}
\keyword{datasets}
