% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{update_par_names}
\alias{update_par_names}
\title{Set model parameter names}
\usage{
update_par_names(nb, par_names)
}
\arguments{
\item{nb}{integer number of unique values FOI can have}

\item{par_names}{vector with model specific parameter names}
}
\value{
param_info_list a list with three items:
nparam integer number of parameters excluding td-foi
nparamtot integer total number of parameters including td-foi
parnames complete list of all parameter names
}
\description{
Based on the model and number of values FOI can have set parameter names
}
\keyword{internal}
