% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_struct.R
\name{write_par_list_yaml}
\alias{write_par_list_yaml}
\title{Write parameter list to YAML file.}
\usage{
write_par_list_yaml(par_list = NULL, file_path = NULL)
}
\arguments{
\item{par_list}{Parameter list—usually the output of PROF::init_par_list}

\item{file_path}{Full path where file to be written.}
}
\description{
Write parameter list to YAML file.
}
