% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download_format.R
\name{csv_to_prof}
\alias{csv_to_prof}
\title{Read a User provided incidence file and populate the PROF data structure}
\usage{
csv_to_prof(filepath, population, location = "PROFVille")
}
\arguments{
\item{filepath}{string.  Path to csv data file}

\item{population}{numeric.  Population size  (no default)}

\item{location}{string. Location name (default is "PROFVille")}
}
\value{
PROF data structure
}
\description{
Given a path to a User provided data file, population size, and
location name populate the PROF data structure.
Data must contain the following fields:
date (%Y-%m-%d format), disease (string, covid19, influenza), metric (string, 'hosp'),
value (numeric, incidence)
}
