% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{est_I0}
\alias{est_I0}
\title{Estimate the initial number of infections}
\usage{
est_I0(inc, mydisease, cadence)
}
\arguments{
\item{inc}{time series of hospitalization incidence}

\item{mydisease}{disease name}

\item{cadence}{integer either 1 or 7 days}
}
\value{
integer estimate for initial number of cases
}
\description{
Use the EpiEstim package to estimate initial number of infections
}
\keyword{internal}
