% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_data}
\alias{fit_data}
\title{Fit model(s) to dataset(s)}
\usage{
fit_data(prof_data = NULL, par_list = NULL)
}
\arguments{
\item{prof_data}{data structure for diseases
for each disease it includes:
disease - covid19 or influenza
population - population size
loc_name - location name
inc_type - incidence type, e.g. hosp_admits
data itself as a 2D structure of dates and incidence}

\item{par_list}{data structure for parameters
for each disease it includes
model - 'seirh' or 'sirh'
constant_dis_pars - initial values for the compartments of the model
dis_par_ranges - a list with parameter ranges and starting values
mcmc_pars}
}
\value{
tab_list - posterior distributionof parameters for each disease
state0 - the initial conditions for all the compartments
wl - the parameter used for tanh of beta(t), default 3
}
\description{
Main driver function for the sequential fitting procedure of one or two
diseases
}
