% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download_format.R
\name{hhs_data_ex}
\alias{hhs_data_ex}
\title{Example use of HHS data functions.}
\usage{
hhs_data_ex(season = NULL, state = "CA", fit_end = NULL)
}
\arguments{
\item{state}{chr. Two-letter state/territory abbreviation.}
}
\value{
PROF data list.
}
\description{
This is a set of HHS function calls that demonstrate how to retrieve and
structure data for use with PROF fitting routines.  This function has
many inputs hard-coded (dates, population, etc) and should be used as a
blueprint rather than as an operational function.
}
\examples{
# download and format HHS hospitalization data for flu and COVID
PROF_data = hhs_data_ex(state="CA")
# plot the contents of a PROF data structure
plot_prof_data(prof_data=PROF_data)
}
