% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot the Results of mechanistic fit model(s) to dataset(s)}
\usage{
plot_fit(prof_data, par_list, fit_list, ntraj = 1000, filename = NULL)
}
\arguments{
\item{prof_data}{data structure for diseases
for each disease it includes:
disease - covid19 or influenza
population - population size
loc_name - location name
inc_type - incidence type, e.g. hosp_admits
data - all available data as a 2D structure of dates and incidence
data_fit - subset of data for fitting (can be eqaul to data)}

\item{par_list}{data structure for parameters
for each disease it includes
model
constant_dis_pars
dis_par_ranges
mcmc_pars}

\item{fit_list}{- output of fit_data. A list with the following components

 tab_list - a list with the posterior distribution(s) for disease(s) fit(s)
 for each disease it includes the population, model parameters and LLK
 state0_list - A list with initial number of individuals in each compartment
 for each pathogen
 wl real, needed for tanh function of bete(t)
 @filename - if NULL print plot to screen, if not also save to filename
 @param ntraj - integer number of stochastic trajectories, default 1000}
}
\value{
plots to screen or file and a list with the following items

fit_traj - a list for each disease containing: model fit mechanistic trajectories,
dates, and reported incidence

pl - a list of ggplot2 objects one for each disease for the mechanistic plots
}
\description{
For the Mechanistic Results, uses the posterior distribution of the fit(s) and
a stochastic code to generate trajectories and calculate
the statistics
}
