% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plots Forecasts for Mechanistic Models}
\usage{
plot_forecast(
  prof_data,
  par_list,
  fit_list,
  ntraj = 1000,
  nfrcst = 35,
  filename = NULL
)
}
\arguments{
\item{prof_data}{data structure for diseases
for each disease it includes:
disease - covid19 or influenza
population - population size
loc_name - location name
inc_type - incidence type, e.g. hosp_admits
data - all available data as a 2D structure of dates and incidence
data_fit - subset of data for fitting (can be eqaul to data)}

\item{par_list}{data structure for parameters
for each disease it includes
model
constant_dis_pars
dis_par_ranges
mcmc_pars}

\item{fit_list}{- output of fit_data. A list with the following components
tab_list - a list with the posterior distribution(s) for disease(s) fit(s)
for each disease it includes the population, model parameters and LLK
state0_list - A list with initial number of individuals in each compartment
for each pathogen
wl real, needed for tanh function of bete(t)
@param ntraj - integer number of stochastic trajectories, default 1000
@param nfrcst - number of time units to produce a forecast for (assume to be the same
cadence as the data), default is 35 days
@filename - if NULL print plot to screen, if not save also to filename}
}
\value{
plots to screen and  file
forecast_traj - a list with a list for each disease and for the combined burden
(random and ordered): trajectories,
dates, and reported incidence.
}
\description{
Uses the posterior distribution of the fit(s) and
a stochastic code to generate trajectories, with the forecast horizon determined
by the call and calculate the statistics
}
