% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stat_forecast.R
\name{plot_stat_forecast}
\alias{plot_stat_forecast}
\title{Baseline statistical forecast}
\usage{
plot_stat_forecast(
  prof_data,
  ntraj = NULL,
  nfrcst = NULL,
  filename = NULL,
  err_cor = 1,
  method_name = "semi_sorted_randA"
)
}
\arguments{
\item{prof_data}{data structure for diseases
for each disease it includes:
disease - covid19 or influenza
population - population size
loc_name - location name
inc_type - incidence type, e.g. hosp_admits
data - all available data as a 2D structure of dates and incidence
data_fit - subset of data for fitting (can be eqaul to data)}

\item{ntraj}{- integer number of stochastic trajectories, default 1000}

\item{nfrcst}{- number of time units to produce a forecast for (assume to be the same
cadence as the data), default is 35 days}

\item{filename}{- if NULL print plot to screen, if not save also to filename}

\item{err_cor}{numeric scalar [-1, 1] - specify the error correlation to be used
for combining influenza and COVID19 forecasts.  If NULL, default behavior is
to evaluate for err_cor=1 (ordered) and err_cor=0 (random).}

\item{method_name}{character - method used to aggregate pathogen forecasts 
 'semi_sorted_randA' or 'lin_scale'. 'semi_sorted_randA' 
is a more proper combination of profiles.  'lin_scale' uses the uncertainty 
range of 'semi_sorted_randA', but ensures that the combined forecast has a 
median equal to the sum of the aggregate forecast medians.}
}
\value{
plots to screen and  file
forecast_traj - a list with a list for each disease and for the combined burden
(random and ordered): trajectories,
dates, and reported incidence.
total_list - a list with the data frames used for each of the plots, these include the median and quantiles
wis_df - a list of data frame with the WIS score of the forecast if available.  If not it contains a NULL
}
\description{
Calculates and plots a baseline statistical forecast for to the dataset(s)
}
